Java Assembler
It can correctly create the intermediate file for basic.txt

It can also mostly create the correct intermediate file for
control_section.txt (the only address incorrect atm is the EQU line).

To test it out, just run the program and then check the
intermediateFile.txt in the "Assembler" folder.